updater/requirements.txt
pm2==0.0.3
GitPython==3.1.43